backend/requirements.txt
aiofiles==23.1.0
alembic==1.11.1
annotated-types==0.5.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bitstring==3.1.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cmake==3.27.4.1
cryptography==38.0.3
DAWG-Python==0.7.2
distlib==0.3.7
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
esptool==4.4
fastapi==0.95.2
fastapi-filter==0.6.1
filelock==3.12.2
fsspec==2023.9.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.16.4
idna==3.4
#jamspell==0.0.12
Jinja2==3.1.2
joblib==1.3.2
lit==16.0.6
Mako==1.2.4
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pipenv==2023.7.11
platformdirs==3.9.1
psycopg==3.1.10
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.7
pydantic_core==2.6.1
PyJWT==2.7.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
redis==4.5.4
reedsolo==1.5.4
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
starlette-context==0.3.6
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.6.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.24.1
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==11.0.3
